[logiclayer] moves geoservice helpers to their own file
