+++ tesseract-server/src/handlers/logic_layer/mod.rs
@@ -1,10 +1,15 @@
 mod aggregate;
+mod geoservice;
 
 pub use self::aggregate::logic_layer_handler;
 pub use self::aggregate::logic_layer_default_handler;
+pub use self::geoservice::GeoserviceQuery;
+pub use self::geoservice::GeoServiceResponseJson;
+pub use self::geoservice::query_geoservice;
 
 use actix_web::{HttpRequest, HttpResponse, Path};
 use crate::app::AppState;
+
 
 pub fn logic_layer_non_unique_levels_default_handler(
     (_req, _cube): (HttpRequest<AppState>, Path<()>),
@@ -12,6 +17,7 @@ pub fn logic_layer_non_unique_levels_default_handler(
 {
     HttpResponse::InternalServerError().body("Error Code 555")
 }
+
 
 pub fn logic_layer_non_unique_levels_handler(
     (_req, _cube): (HttpRequest<AppState>, Path<(String)>),
